+++ web/documentserver-example/csharp-mvc/Helpers/Users.cs
@@ -67,7 +67,9 @@ namespace OnlineEditorsExampleMVC.Helpers
             "Can't mention others in comments",
             "Can't create new files from the editor",
             "Can’t see anyone’s information",
-            "Can't rename files from the editor"
+            "Can't rename files from the editor",
+            "Can't view chat",
+            "View file without collaboration",
         };
 
         private static List<User> users = new List<User>() {