+++ web/documentserver-example/csharp-mvc/Models/FileModel.cs
@@ -150,6 +150,7 @@ namespace OnlineEditorsExampleMVC.Models
                                             { "modifyFilter", editorsMode != "filter" },
                                             { "modifyContentControl", editorsMode != "blockcontent" },
                                             { "review", canEdit && (editorsMode == "edit" || editorsMode == "review") },
+                                            { "chat", !user.id.Equals("uid-0") },
                                             { "reviewGroups", user.reviewGroups },
                                             { "commentGroups", user.commentGroups },
                                             { "userInfoGroups", user.userInfoGroups }
@@ -164,6 +165,11 @@ namespace OnlineEditorsExampleMVC.Models
                                 { "mode", mode },
                                 { "lang", request.Cookies.GetOrDefault("ulang", "en") },
                                 { "callbackUrl", CallbackUrl },  // absolute URL to the document storage service
+                                { "coEditing", editorsMode == "view" && user.id.Equals("uid-0") ? 
+                                    new Dictionary<string, object>{
+                                        {"mode", "strict"},
+                                        {"change", false}
+                                    } : null },
                                 { "createUrl", !user.id.Equals("uid-0") ? createUrl : null },
                                 { "templates", user.templates ? templates : null },
                                 {
@@ -197,7 +203,7 @@ namespace OnlineEditorsExampleMVC.Models
                                             {
                                                 "goback", new Dictionary<string, object>  // settings for the Open file location menu button and upper right corner button
                                                     {
-                                                        { "url", url.Action("Index", "Home") }  // the absolute URL to the website address which will be opened when clicking the Open file location menu button
+                                                        { "url", DocManagerHelper.GetServerUrl(false) }  // the absolute URL to the website address which will be opened when clicking the Open file location menu button
                                                     }
                                             }
                                         }