+++ web/documentserver-example/csharp/DocEditor.aspx.cs
@@ -81,7 +81,7 @@ namespace OnlineEditorsExample
                     + "webeditor.ashx";
                 callbackUrl.Query = "type=track"
                                     + "&fileName=" + HttpUtility.UrlEncode(FileName)
-                                    + "&userAddress=" + HttpUtility.UrlEncode(HttpContext.Current.Request.UserHostAddress);
+                                    + "&userAddress=" + HttpUtility.UrlEncode(_Default.CurUserHostAddress(HttpContext.Current.Request.UserHostAddress));
                 return callbackUrl.ToString();
             }
         }
@@ -102,7 +102,7 @@ namespace OnlineEditorsExample
         // get url to download a file
         public static string getDownloadUrl(string fileName, Boolean isServer = true)
         {
-            var userAddress = isServer ? "&userAddress=" + HttpUtility.UrlEncode(HttpContext.Current.Request.UserHostAddress) : "";
+            var userAddress = isServer ? "&userAddress=" + HttpUtility.UrlEncode(_Default.CurUserHostAddress(HttpContext.Current.Request.UserHostAddress)) : "";
             var downloadUrl = new UriBuilder(_Default.GetServerUrl(isServer));
                 downloadUrl.Path =
                     HttpRuntime.AppDomainAppVirtualPath
@@ -219,6 +219,7 @@ namespace OnlineEditorsExample
                                             { "modifyFilter", editorsMode != "filter" },
                                             { "modifyContentControl", editorsMode != "blockcontent" },
                                             { "review", canEdit && (editorsMode == "edit" || editorsMode == "review") },
+                                            { "chat", !user.id.Equals("uid-0") },
                                             { "reviewGroups", user.reviewGroups },
                                             { "commentGroups", user.commentGroups },
                                             { "userInfoGroups", user.userInfoGroups }
@@ -233,6 +234,11 @@ namespace OnlineEditorsExample
                                 { "mode", mode },
                                 { "lang", Request.Cookies.GetOrDefault("ulang", "en") },
                                 { "callbackUrl", CallbackUrl },  // absolute URL to the document storage service
+                                { "coEditing", editorsMode == "view" && user.id.Equals("uid-0") ? 
+                                    new Dictionary<string, object>{
+                                        {"mode", "strict"},
+                                        {"change", false}
+                                    } : null },
                                 { "createUrl", !user.id.Equals("uid-0") ? createUrl : null },
                                 { "templates", user.templates ? templates : null },
                                 {
@@ -551,7 +557,7 @@ namespace OnlineEditorsExample
         // create the public history url
         private string MakePublicHistoryUrl(string filename, string version, string file, Boolean isServer = true)
         {
-            var userAddress = isServer ? "&userAddress=" + HttpUtility.UrlEncode(HttpContext.Current.Request.UserHostAddress) : "";
+            var userAddress = isServer ? "&userAddress=" + HttpUtility.UrlEncode(_Default.CurUserHostAddress(HttpContext.Current.Request.UserHostAddress)) : "";
             var fileUrl = new UriBuilder(_Default.GetServerUrl(isServer));
             fileUrl.Path = HttpRuntime.AppDomainAppVirtualPath
                 + (HttpRuntime.AppDomainAppVirtualPath.EndsWith("/") ? "" : "/")