+++ web/documentserver-example/java-spring/src/main/java/com/onlyoffice/integration/ExampleData.java
@@ -42,7 +42,9 @@ public class ExampleData {
                 "Can't mention others in comments",
                 "Can't create new files from the editor",
                 "Can’t see anyone’s information",
-                "Can't rename files from the editor"
+                "Can't rename files from the editor",
+                "Can't view chat",
+                "View file without collaboration"
         );
         List<String> description_user_1 = List.of(  // the description for user 1
                 "File author by default",
@@ -51,7 +53,8 @@ public class ExampleData {
                 "He can do everything with the comments",
                 "The file favorite state is undefined",
                 "Can create a file from a template with data from the editor",
-                "Can see the information about all users"
+                "Can see the information about all users",
+                "Can view chat"
         );
         List<String> description_user_2 = List.of(  // the description for user 2
                 "He belongs to Group2",
@@ -59,7 +62,8 @@ public class ExampleData {
                 "He can view every comment, edit his comments and the comments left by the users who don't belong to any of the groups and remove only his comments",
                 "This file is favorite",
                 "Can create a file from an editor",
-                "Can see the information about users from Group2 and users who don’t belong to any group"
+                "Can see the information about users from Group2 and users who don’t belong to any group",
+                "Can view chat"
         );
         List<String> description_user_3 = List.of(  // the description for user 3
                 "He belongs to Group3",
@@ -70,7 +74,8 @@ public class ExampleData {
                 "He can’t download the file",
                 "He can’t print the file",
                 "Can create a file from an editor",
-                "Can see the information about Group2 users"
+                "Can see the information about Group2 users",
+                "Can view chat"
         );
         userService.createUser("John Smith", "smith@example.com",  // create user 1 with the specified parameters
                 description_user_1, "", List.of(FilterState.NULL.toString()),
@@ -78,15 +83,18 @@ public class ExampleData {
                 List.of(FilterState.NULL.toString()),
                 List.of(FilterState.NULL.toString()),
                 List.of(FilterState.NULL.toString()),
-                null);
+                null, true);
         userService.createUser("Mark Pottato", "pottato@example.com",  // create user 2 with the specified parameters
                 description_user_2, "group-2", List.of("","group-2"), List.of(FilterState.NULL.toString()),
-                List.of("group-2", ""), List.of("group-2"), List.of("group-2", ""), true);
+                List.of("group-2", ""), List.of("group-2"), List.of("group-2", ""), true,
+                true);
         userService.createUser("Hamish Mitchell", "mitchell@example.com",  // create user 3 with the specified parameters
                 description_user_3, "group-3", List.of("group-2"), List.of("group-2", "group-3"),
-                List.of("group-2"), new ArrayList<>(), List.of("group-2"), false);
+                List.of("group-2"), new ArrayList<>(), List.of("group-2"), false,
+                true);
         userService.createUser("Anonymous",null,  // create user 0 with the specified parameters
                 description_user_0,"", List.of(FilterState.NULL.toString()), List.of(FilterState.NULL.toString()),
-                List.of(FilterState.NULL.toString()), List.of(FilterState.NULL.toString()), new ArrayList<>(),null);
+                List.of(FilterState.NULL.toString()), List.of(FilterState.NULL.toString()), new ArrayList<>(),null,
+                false);
     }
 }