+++ web/documentserver-example/java-spring/src/main/java/com/onlyoffice/integration/documentserver/models/enums/Action.java
@@ -25,6 +25,7 @@ public enum Action {
     embedded,
     filter,
     comment,
+    chat,
     fillForms,
     blockcontent
 }