+++ web/documentserver-example/java/src/main/java/entities/FileModel.java
@@ -78,6 +78,12 @@ public class FileModel
         // set the editor config parameters
         editorConfig = new EditorConfig(actionData);
         editorConfig.callbackUrl = DocumentManager.GetCallback(fileName);  // get callback url
+
+        editorConfig.coEditing = mode.equals("view") && user.id.equals("uid-0") ? 
+        new HashMap<String, Object>()  {{ 
+            put("mode", "strict");
+            put("change", false);
+        }} : null;
         
         if (lang != null) editorConfig.lang = lang;  // write language parameter to the config
 
@@ -268,9 +274,11 @@ public class FileModel
         public Boolean modifyFilter;
         public Boolean modifyContentControl;
         public Boolean review;
+        public Boolean chat;
         public List<String> reviewGroups;
         public CommentGroups commentGroups;
         public List<String> userInfoGroups;
+        //public Gson gson = new Gson();
 
         // defines what can be done with a document
         public Permissions(String mode, String type, Boolean canEdit, User user)
@@ -284,6 +292,7 @@ public class FileModel
             modifyFilter = !mode.equals("filter");
             modifyContentControl = !mode.equals("blockcontent");
             review = canEdit && (mode.equals("edit") || mode.equals("review"));
+            chat = !user.id.equals("uid-0");
             reviewGroups = user.reviewGroups;
             commentGroups = user.commentGroups;
             userInfoGroups = user.userInfoGroups;
@@ -308,6 +317,7 @@ public class FileModel
         public HashMap<String, Object> actionLink = null;
         public String mode = "edit";
         public String callbackUrl;
+        public HashMap<String, Object> coEditing = null;
         public String lang = "en";
         public String createUrl;
         public List<Map<String, String>> templates;