+++ web/documentserver-example/java/src/main/java/helpers/Users.java
@@ -65,6 +65,8 @@ public class Users {
         add("Can't create new files from the editor");
         add("Can’t see anyone’s information");
         add("Can't rename files from the editor");
+        add("Can't view chat");
+        add("View file without collaboration");
     }};
 
     private static List<User> users = new ArrayList<User>() {{