+++ web/documentserver-example/php/doceditor.php
@@ -108,6 +108,7 @@
                 "modifyFilter" => $editorsMode != "filter",
                 "modifyContentControl" => $editorsMode != "blockcontent",
                 "review" => $canEdit && ($editorsMode == "edit" || $editorsMode == "review"),
+                "chat" => $user->id != "uid-0",
                 "reviewGroups" => $user->reviewGroups,
                 "commentGroups" => $user->commentGroups,
                 "userInfoGroups" => $user->userInfoGroups
@@ -118,6 +119,10 @@
             "mode" => $mode,
             "lang" => empty($_COOKIE["ulang"]) ? "en" : $_COOKIE["ulang"],
             "callbackUrl" => getCallbackUrl($filename),  // absolute URL to the document storage service
+            "coEditing" => $editorsMode == "view" && $user->id == "uid-0" ? [
+                "mode" => "strict", 
+                "change" => false
+            ] : null,
             "createUrl" => $user->id != "uid-0" ? $createUrl : null,
             "templates" => $user->templates ? $templates : null,
             "user" => [  // the user currently viewing or editing the document